ATR-917: merged PX1013 analysis into the general analyzer LongOperationInGraphAnalyzer for long running operations in graphs and graph extensions
