+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/LongOperationStart/ActionHandlerReturnTypeFix.cs
@@ -1,5 +1,4 @@
-﻿
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
 using System.Composition;
 using System.Threading.Tasks;
 
@@ -9,7 +8,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace Acuminator.Analyzers.StaticAnalysis.ActionHandlerReturnType
+namespace Acuminator.Analyzers.StaticAnalysis.LongOperationStart
 {
 	[ExportCodeFixProvider(LanguageNames.CSharp), Shared]
 	public class ActionHandlerReturnTypeFix : PXCodeFixProvider
@@ -17,23 +16,15 @@ namespace Acuminator.Analyzers.StaticAnalysis.ActionHandlerReturnType
 		public override ImmutableArray<string> FixableDiagnosticIds { get; } =
 			ImmutableArray.Create(Descriptors.PX1013_PXActionHandlerInvalidReturnType.Id);
 
-        protected override async Task RegisterCodeFixesForDiagnosticAsync(CodeFixContext context, Diagnostic diagnostic)
+		protected override Task RegisterCodeFixesForDiagnosticAsync(CodeFixContext context, Diagnostic diagnostic)
 		{
 			context.CancellationToken.ThrowIfCancellationRequested();
 
-			var root = await context.Document
-				.GetSyntaxRootAsync(context.CancellationToken)
-				.ConfigureAwait(false);
-
-			if (root?.FindNode(context.Span) is not MethodDeclarationSyntax node)
-			{
-				return;
-			}
-
 			var changeReturnTypeTitle = nameof(Resources.PX1013Fix).GetLocalized().ToString();
-			var codeFixAction = new InvalidPXActionSignatureFix.ChangeSignatureAction(changeReturnTypeTitle, context.Document, node);
+			var codeFixAction = new InvalidPXActionSignatureFix.ChangeSignatureAction(changeReturnTypeTitle, context.Document, context.Span);
 
 			context.RegisterCodeFix(codeFixAction, diagnostic);
+			return Task.CompletedTask;
 		}
 	}
 }
